requirements.txt
streamlit
pandas
python-dotenv
requests
pyarrow==15.0.0